@@ -5,16 +5,18 @@
   "proxy": "http://localhost:3001",
   "dependencies": {
     "@apollo/client": "^3.6.9",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^4.4.1",
-    "jwt-decode": "^2.2.0",
-    "react": "^16.13.1",
+    "graphql": "^16.5.0",
+    "jwt-decode": "^3.1.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^1.0.1",
-    "react-dom": "^16.13.1",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.1"
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts --openssl-legacy-provider start",
