@@ -7,42 +7,35 @@ import { useMutation } from '@apollo/client';
 import { LOGIN_USER } from '../utils/mutations';
 
 const LoginForm = () => {
-  const [login, { error }] = useMutation(LOGIN_USER);
-
-  const [userFormData, setUserFormData] = useState({ email: '', password: '' });
   const [validated] = useState(false);
   const [showAlert, setShowAlert] = useState(false);
 
-  const handleInputChange = (event) => {
-    const { name, value } = event.target;
-    setUserFormData({ ...userFormData, [name]: value });
-  };
+  const [login, { error }] = useMutation(LOGIN_USER);
 
+  // submit form
   const handleFormSubmit = async (event) => {
     event.preventDefault();
 
-    // check if form has everything (as per react-bootstrap docs)
-    const form = event.currentTarget;
-    if (form.checkValidity() === false) {
-      event.preventDefault();
-      event.stopPropagation();
-    }
-
     try {
       const { data } = await login({
-        variables: { ...userFormData },
+        variables: { ...formState },
       });
 
       Auth.login(data.login.token);
-    } catch (err) {
-      console.error(err);
-      setShowAlert(true);
+    } catch (e) {
+      console.error(e);
     }
+  };
 
-    setUserFormData({
-      username: '',
-      email: '',
-      password: '',
+  const [formState, setFormState] = useState({ email: '', password: '' });
+
+  // update state based on form input changes
+  const handleChange = (event) => {
+    const { name, value } = event.target;
+
+    setFormState({
+      ...formState,
+      [name]: value,
     });
   };
 
@@ -63,8 +56,8 @@ const LoginForm = () => {
             type="text"
             placeholder="Your email"
             name="email"
-            onChange={handleInputChange}
-            value={userFormData.email}
+            value={formState.email}
+            onChange={handleChange}
             required
           />
           <Form.Control.Feedback type="invalid">
@@ -78,8 +71,8 @@ const LoginForm = () => {
             type="password"
             placeholder="Your password"
             name="password"
-            onChange={handleInputChange}
-            value={userFormData.password}
+            value={formState.password}
+            onChange={handleChange}
             required
           />
           <Form.Control.Feedback type="invalid">
@@ -87,7 +80,7 @@ const LoginForm = () => {
           </Form.Control.Feedback>
         </Form.Group>
         <Button
-          disabled={!(userFormData.email && userFormData.password)}
+          disabled={!(formState.email && formState.password)}
           type="submit"
           variant="success"
         >
