@@ -1,43 +1,47 @@
-// use this to decode a token and get the user's information out of it
 import decode from 'jwt-decode';
 
-// create a new class to instantiate for a user
 class AuthService {
-  // get user data
+  // retrieve data saved in token
   getProfile() {
     return decode(this.getToken());
   }
 
-  // check if user's logged in
   loggedIn() {
-    // Checks if there is a saved token and it's still valid
-    const token = this.getToken();
-    return !!token && !this.isTokenExpired(token); // handwaiving here
+    // check if the user is still logged in
+    const token = this.getToken(); // Checks if there is a saved token and it's still valid
+    // use type coersion to check if token is NOT undefined and the token is NOT expired
+    return !!token && !this.isTokenExpired(token);
   }
 
-  // check if token is expired
+  // check if the token has expired
   isTokenExpired(token) {
     try {
       const decoded = decode(token);
       if (decoded.exp < Date.now() / 1000) {
         return true;
-      } else return false;
+      } else {
+        return false;
+      }
     } catch (err) {
       return false;
     }
   }
 
+  // retrieve token from localStorage
   getToken() {
     // Retrieves the user token from localStorage
     return localStorage.getItem('id_token');
   }
 
+  // set token to localStorage and reload page to homepage
   login(idToken) {
     // Saves user token to localStorage
     localStorage.setItem('id_token', idToken);
+
     window.location.assign('/');
   }
 
+  // clear token from localStorage and force logout with reload
   logout() {
     // Clear user token and profile data from localStorage
     localStorage.removeItem('id_token');
@@ -45,5 +49,4 @@ class AuthService {
     window.location.assign('/');
   }
 }
-
 export default new AuthService();
