@@ -1,3 +1,4 @@
 const User = require('./User');
+const Book = require('./Book');
 
-module.exports = { User };
+module.exports = { User, Book };
